This change brings in a doozy of a refactor but for the better!
Code handling examples and copying assets now executes at *build time*

Examples directory has its own isolated CMakeLists.txt file now The
`PLUSH_BUILD_EXAMPLES` variable will *not* be available to be set if
SDL2 cannot be found. (When trying to "find" SDL2 in a custom location,
you can pass `-DSDL2_ROOT=/directory/name/containing/include+lib` to
more easily handle custom locations.

Any tool that relies on `compile_command.json` will now "just work"
after a configuration.

Examples can be run directly from any build system. For the easiest
approach, simply run `cmake --build <build-directory> --target
run.<example-name>` as this command works with *all* build systems
(Xcode and Visual Studio included)

Public headers were moved into the `include` directory to make the
`install()` call work correctly for directory layout. `plush.h` and co.
are now isolated inside of a `plush/` directory, and would be included
as `#include <plush/plush.h>`

The `pl_conf.h` header was renamed to `pl_conf.h.in` and converted to a
`configure_file` call (though we could technically use
`file(CONFIGURE)`. The `PL_MAX_CHILDREN`, `PL_MAX_LIGHTS`, and the
`PL_MAX_TRIANGLES` variables are now configurable from CMake, with
default fallbacks to their original settings.

Better `stdint.h` detection. We now check if the current environment
*has* a stdint.h (not ideal but oh well). We then let the compiler use
`__has_include` first, fallback to the `HAVE_STDINT_H` cache variable
value, and then if either of those fail, users can manually define the
path to a PLUSH_STDINT_H header. This entire feature is moderately
fragile and as I'm typing this paragraph I'm realizing the mistakes I
made implementing it at 2 A.M. local.

A large amount of logic was moved to generator expressions. Generator
expressions are powerful and let you do inline logic at generation time,
which executes in parallel, versus the single threaded nature of the
configure step.
