+++ examples/ex4.c
@@ -8,7 +8,7 @@
 #include <math.h>
 
 // Include the plush header file
-#include <plush.h> 
+#include <plush/plush.h> 
 
 // Include our example graphics interface module
 #include "ex.h"
